requirements.txt
streamlit>=1.37.0
python-dotenv>=1.0.1
anthropic>=0.33.0
openai>=1.37.0
pinecone>=5.0.0
cohere>=5.5.0
